7 hours of ui learning today
Made basic ui, schema and template, now working it into code and to my liking. after that start connecting it to database and work on setting up the right cascade in the database aswell so i can test around with ading and deleting loadouts. then can stat working on combat module and controllers.