Database changes! costpershot function
Alright here is a short list of changes:
Added tt value to enhancers used only for calculations.
changed decay to a Decimal instead of  float for more precise calculations.
added a getCostPerShot function to the loadoutManager. this gets the total cost of the loadout, weapon+ al lattachemnt + enhancer decay + ammo burn, also takes into account enhancers breaking at a rate of 1/2500 as per the default of the games wiki in the weaponcompareV2