Big changes again, ORM, Sessions
discovered something from the orm today, Scoped sessions. which is very useful for my application, saves on some double query's and is more thread save, had to, and probably have to change some of my functions due too it. ad have to remember to commit myself as scoped_session cannot use context manager protocol.

but now same session is used everywhere in the main thread which helps with keeping relations of queried objects (lazy loaded) so I wont have to do joinedloads/"manual" joins as much

also restructuring my db file as the db file is the module that interacts with the database (db level functions).

which first had a class called DB, but as in python the file itself is treated as a module when imported (so kind of  acts as an class)  i dont have to make a class in it.

it still exists as of now but is to be removed in the next commit.

Also split of the logginrun_bll class into its own file without class. as it was first shared in the combat module file