small changes, progress combat module
TODO check where and how to add new data to database, I kinda wanted to so a seperate save function that runs every 10-15 sec on a seperate thread but maybe I might as well do it on the end of every tick, as not to many lines are added at once. and as it is local it shouldnt hold the rest, i will try this first and if it does end up breaking. im gonn ahev to remake / think some tings