combatmodule addition
cleaned up combat module a bit more. almost to a point where I could start testing it, still have to figure out if I want to define the modules in the base app or in the main frame, I personally think in the app would be better approach.

Now to a point where I need to start doing more UI code to be able to really test the logger and module. I was thinking of making a tabbed window inside the main program for switching between the main controls, table/graph data, more tba.

and then also make a view for every tab, so main priority right now is implementing the tabbed layout in the main view (main window) and then make a temporary view that can start and stop the logger.

if all that works my next goal is configuration and settings control so I can start working with less hardcoded data.

then most likely a logger for the program itself to make debugging easier.